Accounting for default case and Commentary Functions
The default case is now accounted for and is appropriately handled. Additionally, the functionality for writing the changes to the file and to a new file has been added and updated to work correctly and fully. The commentary function generating function comments and has been revamped, it is all working apart from the tab spacing.