Starting on Menu Builder
The menu builder is being remade to be more modular and account for different styles of programming and is currently in the process of being created. The reading back of the template layout is working and the locating of the start and end of the scope for the Main Menu Loop section of the code has been tested and works.